Add Playback Service
