+++ MiniSoundsChallenge/ViewModels/RailViewModel.swift
@@ -8,12 +8,14 @@
 import Foundation
 
 class RailViewModel: ObservableObject {
+    let playbackService: PlaybackService
     var title: String
     @Published var stations: [Stations.Module.StationData]
     
-    init(title: String, stations: [Stations.Module.StationData]) {
+    init(title: String, stations: [Stations.Module.StationData], playbackService: PlaybackService) {
         self.title = title
         self.stations = stations
+        self.playbackService = playbackService
     }
     
 }
